HPCC-23548 Clean up various uses for the environment
