+++ common/workunit/workunit.cpp
@@ -4571,20 +4571,6 @@ private:
 
 //==========================================================================================
 
-class CStringArrayIterator : implements IStringIterator, public CInterface
-{
-    unsigned idx;
-    StringArray strings;
-public:
-    IMPLEMENT_IINTERFACE;
-    CStringArrayIterator() { idx = 0; };
-    void append(const char *str) { strings.append(str); }
-    virtual bool first() { idx = 0; return strings.isItem(idx); }
-    virtual bool next() { idx ++; return strings.isItem(idx); }
-    virtual bool isValid() { return strings.isItem(idx); }
-    virtual IStringVal & str(IStringVal &s) { s.set(strings.item(idx)); return s; }
-};
-
 class CCachedJobNameIterator : implements IStringIterator, public CInterface
 {
     Owned<IPropertyTreeIterator> it;
@@ -4595,16 +4581,6 @@ public:
     virtual bool next() { return it->next(); }
     virtual bool isValid() { return it->isValid(); }
     virtual IStringVal & str(IStringVal &s) { s.set(it->query().queryName()+1); return s; }
-};
-
-class CEmptyStringIterator : implements IStringIterator, public CInterface
-{
-public:
-    IMPLEMENT_IINTERFACE;
-    virtual bool first() { return false; }
-    virtual bool next() { return false; }
-    virtual bool isValid() { return false; }
-    virtual IStringVal & str(IStringVal &s) { s.clear(); return s; }
 };
 
 EnumMapping workunitSortFields[] =
@@ -7299,643 +7275,6 @@ wuTokenStates verifyWorkunitDAToken(const char * ctxUser, const char * daToken)
 
     return wuActive ? wuTokenValid : wuTokenWorkunitInactive;
 }
-
-IPropertyTree *queryRoxieProcessTree(IPropertyTree *environment, const char *process)
-{
-    if (!process || !*process)
-        return NULL;
-    VStringBuffer xpath("Software/RoxieCluster[@name=\"%s\"]", process);
-    return environment->queryPropTree(xpath.str());
-}
-
-void getRoxieProcessServers(IPropertyTree *roxie, SocketEndpointArray &endpoints)
-{
-    if (!roxie)
-        return;
-    Owned<IPropertyTreeIterator> servers = roxie->getElements("RoxieServerProcess");
-    ForEach(*servers)
-    {
-        IPropertyTree &server = servers->query();
-        const char *netAddress = server.queryProp("@netAddress");
-        if (netAddress && *netAddress)
-        {
-            SocketEndpoint ep(netAddress, server.getPropInt("@port", 9876));
-            endpoints.append(ep);
-        }
-    }
-}
-
-void getRoxieProcessServers(const char *process, SocketEndpointArray &servers)
-{
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    getRoxieProcessServers(queryRoxieProcessTree(root, process), servers);
-}
-
-class CEnvironmentClusterInfo: implements IConstWUClusterInfo, public CInterface
-{
-    StringAttr name;
-    StringAttr alias;
-    StringAttr serverQueue;
-    StringAttr agentQueue;
-    StringAttr agentName;
-    StringAttr roxieProcess;
-    SocketEndpointArray roxieServers;
-    StringAttr thorQueue;
-    StringArray thorProcesses;
-    StringArray primaryThorProcesses;
-    StringAttr prefix;
-    StringAttr ldapUser;
-    StringBuffer ldapPassword;
-    ClusterType platform;
-    unsigned clusterWidth;
-    unsigned roxieRedundancy;
-    unsigned channelsPerNode;
-    unsigned numberOfSlaveLogs;
-    int roxieReplicateOffset;
-
-public:
-    IMPLEMENT_IINTERFACE;
-    CEnvironmentClusterInfo(const char *_name, const char *_prefix, const char *_alias, IPropertyTree *agent, IArrayOf<IPropertyTree> &thors, IPropertyTree *roxie)
-        : name(_name), prefix(_prefix), alias(_alias), roxieRedundancy(0), channelsPerNode(0), numberOfSlaveLogs(0), roxieReplicateOffset(1)
-    {
-        StringBuffer queue;
-        if (thors.ordinality())
-        {
-            thorQueue.set(getClusterThorQueueName(queue.clear(), name));
-            clusterWidth = 0;
-            bool isMultiThor = (thors.length() > 1);
-            ForEachItemIn(i,thors) 
-            {
-                IPropertyTree &thor = thors.item(i);
-                const char* thorName = thor.queryProp("@name");
-                thorProcesses.append(thorName);
-                if (!isMultiThor)
-                    primaryThorProcesses.append(thorName);
-                else
-                {
-                    const char *nodeGroup = thor.queryProp("@nodeGroup");
-                    if (!nodeGroup || strieq(nodeGroup, thorName))
-                        primaryThorProcesses.append(thorName);
-                }
-                unsigned nodes = thor.getCount("ThorSlaveProcess");
-                unsigned slavesPerNode = thor.getPropInt("@slavesPerNode", 1);
-                unsigned channelsPerSlave = thor.getPropInt("@channelsPerSlave", 1);
-                unsigned ts = nodes * slavesPerNode * channelsPerSlave;
-                numberOfSlaveLogs = nodes * slavesPerNode;
-                if (clusterWidth && (ts!=clusterWidth)) 
-                    throw MakeStringException(WUERR_MismatchClusterSize,"CEnvironmentClusterInfo: mismatched thor sizes in cluster");
-                clusterWidth = ts;
-            }
-            platform = ThorLCRCluster;
-        }
-        else if (roxie)
-        {
-            roxieProcess.set(roxie->queryProp("@name"));
-            platform = RoxieCluster;
-            getRoxieProcessServers(roxie, roxieServers);
-            clusterWidth = roxieServers.length();
-            ldapUser.set(roxie->queryProp("@ldapUser"));
-            StringBuffer encPassword (roxie->queryProp("@ldapPassword"));
-            if (encPassword.length())
-                decrypt(ldapPassword, encPassword);
-            const char *redundancyMode = roxie->queryProp("@slaveConfig");
-            if (redundancyMode && *redundancyMode)
-            {
-                unsigned dataCopies = roxie->getPropInt("@numDataCopies", 1);
-                if (strieq(redundancyMode, "overloaded"))
-                    channelsPerNode = roxie->getPropInt("@channelsPernode", 1);
-                else if (strieq(redundancyMode, "full redundancy"))
-                {
-                    roxieRedundancy = dataCopies-1;
-                    roxieReplicateOffset = 0;
-                }
-                else if (strieq(redundancyMode, "cyclic redundancy"))
-                {
-                    roxieRedundancy = dataCopies-1;
-                    channelsPerNode = dataCopies;
-                    roxieReplicateOffset = roxie->getPropInt("@cyclicOffset", 1);
-                }
-            }
-        }
-        else 
-        {
-            clusterWidth = 1;
-            platform = HThorCluster;
-        }
-
-        if (agent)
-        {
-            assertex(!roxie);
-            agentQueue.set(getClusterEclAgentQueueName(queue.clear(), name));
-            agentName.set(agent->queryProp("@name"));
-        }
-        else if (roxie)
-            agentQueue.set(getClusterRoxieQueueName(queue.clear(), name));
-        // MORE - does this need to be conditional?
-        serverQueue.set(getClusterEclCCServerQueueName(queue.clear(), name));
-    }
-
-    IStringVal & getName(IStringVal & str) const
-    {
-        str.set(name.get());
-        return str;
-    }
-    const char *getAlias() const
-    {
-        return alias;
-    }
-    IStringVal & getScope(IStringVal & str) const
-    {
-        str.set(prefix.get());
-        return str;
-    }
-    IStringVal & getAgentQueue(IStringVal & str) const
-    {
-        str.set(agentQueue);
-        return str;
-    }
-    IStringVal & getAgentName(IStringVal & str) const
-    {
-        str.set(agentName);
-        return str;
-    }
-    virtual IStringVal & getServerQueue(IStringVal & str) const
-    {
-        str.set(serverQueue);
-        return str;
-    }
-    IStringVal & getThorQueue(IStringVal & str) const
-    {
-        str.set(thorQueue);
-        return str;
-    }
-    unsigned getSize() const 
-    {
-        return clusterWidth;
-    }
-    unsigned getNumberOfSlaveLogs() const
-    {
-        return numberOfSlaveLogs;
-    }
-    virtual ClusterType getPlatform() const
-    {
-        return platform;
-    }
-    IStringVal & getRoxieProcess(IStringVal & str) const
-    {
-        str.set(roxieProcess.get());
-        return str;
-    }
-    const StringArray & getThorProcesses() const
-    {
-        return thorProcesses;
-    }
-    const StringArray & getPrimaryThorProcesses() const
-    {
-        return primaryThorProcesses;
-    }
-
-    const SocketEndpointArray & getRoxieServers() const
-    {
-        return roxieServers;
-    }
-    unsigned getRoxieRedundancy() const
-    {
-        return roxieRedundancy;
-    }
-    unsigned getChannelsPerNode() const
-    {
-        return channelsPerNode;
-    }
-    int getRoxieReplicateOffset() const
-    {
-        return roxieReplicateOffset;
-    }
-    const char *getLdapUser() const
-    {
-        return ldapUser.get();
-    }
-    virtual const char *getLdapPassword() const
-    {
-        return ldapPassword.str();
-    }
-};
-
-IStringVal &getProcessQueueNames(IStringVal &ret, const char *process, const char *type, const char *suffix)
-{
-    if (process)
-    {
-        Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-        Owned<IConstEnvironment> env = factory->openEnvironment();
-        Owned<IPropertyTree> root = &env->getPTree();
-        StringBuffer queueNames;
-        StringBuffer xpath;
-        xpath.appendf("%s[@process=\"%s\"]", type, process);
-        Owned<IPropertyTreeIterator> targets = root->getElements("Software/Topology/Cluster");
-        ForEach(*targets)
-        {
-            IPropertyTree &target = targets->query();
-            if (target.hasProp(xpath))
-            {
-                if (queueNames.length())
-                    queueNames.append(',');
-                queueNames.append(target.queryProp("@name")).append(suffix);
-            }
-        }
-        ret.set(queueNames);
-    }
-    return ret;
-}
-
-#define ROXIE_QUEUE_EXT ".roxie"
-#define THOR_QUEUE_EXT ".thor"
-#define ECLCCSERVER_QUEUE_EXT ".eclserver"
-#define ECLSERVER_QUEUE_EXT ECLCCSERVER_QUEUE_EXT
-#define ECLSCHEDULER_QUEUE_EXT ".eclscheduler"
-#define ECLAGENT_QUEUE_EXT ".agent"
-
-extern WORKUNIT_API void getDFUServerQueueNames(StringArray &ret, const char *process)
-{
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-
-    StringBuffer xpath ("Software/DfuServerProcess");
-    if (!isEmptyString(process))
-        xpath.appendf("[@name=\"%s\"]", process);
-
-    Owned<IPropertyTree> root = &env->getPTree();
-    Owned<IPropertyTreeIterator> targets = root->getElements(xpath.str());
-    ForEach(*targets)
-    {
-        IPropertyTree &target = targets->query();
-        if (target.hasProp("@queue"))
-            ret.appendListUniq(target.queryProp("@queue"), ",");
-    }
-    return;
-}
-
-extern WORKUNIT_API IStringVal &getEclCCServerQueueNames(IStringVal &ret, const char *process)
-{
-    return getProcessQueueNames(ret, process, "EclCCServerProcess", ECLCCSERVER_QUEUE_EXT);
-}
-
-extern WORKUNIT_API IStringVal &getEclServerQueueNames(IStringVal &ret, const char *process)
-{
-    return getProcessQueueNames(ret, process, "EclServerProcess", ECLSERVER_QUEUE_EXT); // shares queue name with EclCCServer
-}
-
-extern WORKUNIT_API IStringVal &getEclSchedulerQueueNames(IStringVal &ret, const char *process)
-{
-    return getProcessQueueNames(ret, process, "EclSchedulerProcess", ECLSCHEDULER_QUEUE_EXT); // Shares deployment/config with EclCCServer
-}
-
-extern WORKUNIT_API IStringVal &getAgentQueueNames(IStringVal &ret, const char *process)
-{
-    return getProcessQueueNames(ret, process, "EclAgentProcess", ECLAGENT_QUEUE_EXT);
-}
-
-extern WORKUNIT_API IStringVal &getRoxieQueueNames(IStringVal &ret, const char *process)
-{
-    return getProcessQueueNames(ret, process, "RoxieCluster", ROXIE_QUEUE_EXT);
-}
-
-extern WORKUNIT_API IStringVal &getThorQueueNames(IStringVal &ret, const char *process)
-{
-    return getProcessQueueNames(ret, process, "ThorCluster", THOR_QUEUE_EXT);
-}
-
-extern WORKUNIT_API StringBuffer &getClusterThorQueueName(StringBuffer &ret, const char *cluster)
-{
-    return ret.append(cluster).append(THOR_QUEUE_EXT);
-}
-
-extern WORKUNIT_API StringBuffer &getClusterThorGroupName(StringBuffer &ret, const char *cluster)
-{
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    StringBuffer path;
-    path.append("Software/ThorCluster[@name=\"").append(cluster).append("\"]");
-    IPropertyTree * child = root->queryPropTree(path);
-    if (child)
-        getClusterGroupName(*child, ret);
-
-    return ret;
-}
-
-extern WORKUNIT_API StringBuffer &getClusterGroupName(StringBuffer &ret, const char *cluster)
-{
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    StringBuffer path;
-    path.set("Software/ThorCluster[@name=\"").append(cluster).append("\"]");
-    IPropertyTree * child = root->queryPropTree(path);
-    if (child)
-    {
-        return getClusterGroupName(*child, ret);
-    }
-    path.set("Software/RoxieCluster[@name=\"").append(cluster).append("\"]");
-    child = root->queryPropTree(path);
-    if (child)
-    {
-        return getClusterGroupName(*child, ret);
-    }
-    path.set("Software/EclAgentProcess[@name=\"").append(cluster).append("\"]");
-    child = root->queryPropTree(path);
-    if (child)
-    {
-        return ret.setf("hthor__%s", cluster);
-    }
-
-    return ret;
-}
-
-extern WORKUNIT_API ClusterType getClusterTypeByClusterName(const char *cluster)
-{
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    StringBuffer path;
-    path.set("Software/ThorCluster[@name=\"").append(cluster).append("\"]");
-    if (root->hasProp(path))
-        return ThorLCRCluster;
-
-    path.set("Software/RoxieCluster[@name=\"").append(cluster).append("\"]");
-    if (root->hasProp(path))
-        return RoxieCluster;
-
-    path.set("Software/EclAgentProcess[@name=\"").append(cluster).append("\"]");
-    if (root->hasProp(path))
-        return HThorCluster;
-
-    return NoCluster;
-}
-
-extern WORKUNIT_API StringBuffer &getClusterRoxieQueueName(StringBuffer &ret, const char *cluster)
-{
-    return ret.append(cluster).append(ROXIE_QUEUE_EXT);
-}
-
-extern WORKUNIT_API StringBuffer &getClusterEclCCServerQueueName(StringBuffer &ret, const char *cluster)
-{
-    return ret.append(cluster).append(ECLCCSERVER_QUEUE_EXT);
-}
-
-extern WORKUNIT_API StringBuffer &getClusterEclServerQueueName(StringBuffer &ret, const char *cluster)
-{
-    return ret.append(cluster).append(ECLSERVER_QUEUE_EXT);
-}
-
-extern WORKUNIT_API StringBuffer &getClusterEclAgentQueueName(StringBuffer &ret, const char *cluster)
-{
-    return ret.append(cluster).append(ECLAGENT_QUEUE_EXT);
-}
-
-extern WORKUNIT_API IStringIterator *getTargetClusters(const char *processType, const char *processName)
-{
-    Owned<CStringArrayIterator> ret = new CStringArrayIterator;
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    StringBuffer xpath;
-    xpath.appendf("%s", processType ? processType : "*");
-    if (processName && *processName)
-        xpath.appendf("[@process=\"%s\"]", processName);
-    Owned<IPropertyTreeIterator> targets = root->getElements("Software/Topology/Cluster");
-    ForEach(*targets)
-    {
-        IPropertyTree &target = targets->query();
-        if (target.hasProp(xpath))
-        {
-            ret->append(target.queryProp("@name"));
-        }
-    }
-    return ret.getClear();
-}
-
-extern WORKUNIT_API bool isProcessCluster(const char *process)
-{
-    if (!process || !*process)
-        return false;
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    VStringBuffer xpath("Software/*Cluster[@name=\"%s\"]", process);
-    return root->hasProp(xpath.str());
-}
-
-extern WORKUNIT_API bool isProcessCluster(const char *remoteDali, const char *process)
-{
-    if (!remoteDali || !*remoteDali)
-        return isProcessCluster(process);
-    if (!process || !*process)
-        return false;
-    Owned<INode> remote = createINode(remoteDali, 7070);
-    if (!remote)
-        return false;
-
-    //Cannot use getEnvironmentFactory() since it is using a remotedali
-    VStringBuffer xpath("Environment/Software/*Cluster[@name=\"%s\"]/@name", process);
-    try
-    {
-        Owned<IPropertyTreeIterator> clusters = querySDS().getElementsRaw(xpath, remote, 1000*60*1);
-        return clusters->first();
-    }
-    catch (IException *E)
-    {
-        StringBuffer msg;
-        E->errorMessage(msg);
-        DBGLOG("Exception validating cluster %s/%s: %s", remoteDali, xpath.str(), msg.str());
-        E->Release();
-    }
-    return true;
-}
-
-IConstWUClusterInfo* getTargetClusterInfo(IPropertyTree *environment, IPropertyTree *cluster)
-{
-    const char *clustname = cluster->queryProp("@name");
-
-    // MORE - at the moment configenv specifies eclagent and thor queues by (in effect) placing an 'example' thor or eclagent in the topology
-    // that uses the queue that will be used.
-    // We should and I hope will change that, at which point the code below gets simpler
-
-    StringBuffer prefix(cluster->queryProp("@prefix"));
-    prefix.toLowerCase();
-
-    StringBuffer xpath;
-    StringBuffer querySetName;
-    
-    IPropertyTree *agent = NULL;
-    const char *agentName = cluster->queryProp("EclAgentProcess/@process");
-    if (agentName) 
-    {
-        xpath.clear().appendf("Software/EclAgentProcess[@name=\"%s\"]", agentName);
-        agent = environment->queryPropTree(xpath.str());
-    }
-    Owned<IPropertyTreeIterator> ti = cluster->getElements("ThorCluster");
-    IArrayOf<IPropertyTree> thors;
-    ForEach(*ti) 
-    {
-        const char *thorName = ti->query().queryProp("@process");
-        if (thorName) 
-        {
-            xpath.clear().appendf("Software/ThorCluster[@name=\"%s\"]", thorName);
-            if (environment->hasProp(xpath.str()))
-                thors.append(*environment->getPropTree(xpath.str()));
-        }
-    }
-    const char *roxieName = cluster->queryProp("RoxieCluster/@process");
-    return new CEnvironmentClusterInfo(clustname, prefix, cluster->queryProp("@alias"), agent, thors, queryRoxieProcessTree(environment, roxieName));
-}
-
-IPropertyTree* getTopologyCluster(Owned<IPropertyTree> &envRoot, const char *clustname)
-{
-    if (!clustname || !*clustname)
-        return NULL;
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    envRoot.setown(&env->getPTree());
-    StringBuffer xpath;
-    xpath.appendf("Software/Topology/Cluster[@name=\"%s\"]", clustname);
-    return envRoot->getPropTree(xpath.str());
-}
-
-bool validateTargetClusterName(const char *clustname)
-{
-    Owned<IPropertyTree> envRoot;
-    Owned<IPropertyTree> cluster = getTopologyCluster(envRoot, clustname);
-    return (cluster.get()!=NULL);
-}
-
-IConstWUClusterInfo* getTargetClusterInfo(const char *clustname)
-{
-    Owned<IPropertyTree> envRoot;
-    Owned<IPropertyTree> cluster = getTopologyCluster(envRoot, clustname);
-    if (!cluster)
-        return NULL;
-    return getTargetClusterInfo(envRoot, cluster);
-}
-
-unsigned getEnvironmentClusterInfo(CConstWUClusterInfoArray &clusters)
-{
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    return getEnvironmentClusterInfo(root, clusters);
-}
-
-unsigned getEnvironmentClusterInfo(IPropertyTree* environmentRoot, CConstWUClusterInfoArray &clusters)
-{
-    if (!environmentRoot)
-        return 0;
-
-    Owned<IPropertyTreeIterator> clusterIter = environmentRoot->getElements("Software/Topology/Cluster");
-    ForEach(*clusterIter)
-    {
-        IPropertyTree &node = clusterIter->query();
-        Owned<IConstWUClusterInfo> cluster = getTargetClusterInfo(environmentRoot, &node);
-        clusters.append(*cluster.getClear());
-    }
-    return clusters.ordinality();
-}
-
-const char *getTargetClusterComponentName(const char *clustname, const char *processType, StringBuffer &name)
-{
-    if (!clustname)
-        return NULL;
-
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    StringBuffer xpath;
-
-    xpath.appendf("Software/Topology/Cluster[@name=\"%s\"]", clustname);
-    Owned<IPropertyTree> cluster = root->getPropTree(xpath.str());
-    if (!cluster) 
-        return NULL;
-
-    StringBuffer xpath1;
-    xpath1.appendf("%s/@process", processType);
-    name.append(cluster->queryProp(xpath1.str()));
-    return name.str();
-}
-
-unsigned getEnvironmentThorClusterNames(StringArray &thorNames, StringArray &groupNames, StringArray &targetNames, StringArray &queueNames)
-{
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    Owned<IPropertyTreeIterator> allTargets = root->getElements("Software/Topology/Cluster");
-    ForEach(*allTargets)
-    {
-        IPropertyTree &target = allTargets->query();
-        const char *targetName = target.queryProp("@name");
-        if (targetName && *targetName)
-        {
-            Owned<IPropertyTreeIterator> thorClusters = target.getElements("ThorCluster");
-            ForEach(*thorClusters)
-            {
-                const char *thorName = thorClusters->query().queryProp("@process");
-                VStringBuffer query("Software/ThorCluster[@name=\"%s\"]",thorName);
-                IPropertyTree *thorCluster = root->queryPropTree(query.str());
-                if (thorCluster)
-                {
-                    const char *groupName = thorCluster->queryProp("@nodeGroup");
-                    if (!groupName||!*groupName)
-                        groupName = thorName;
-                    thorNames.append(thorName);
-                    groupNames.append(groupName);
-                    targetNames.append(targetName);
-                    StringBuffer queueName(targetName);
-                    queueNames.append(queueName.append(THOR_QUEUE_EXT));
-                }
-            }
-        }
-    }
-    return thorNames.ordinality();
-}
-
-
-unsigned getEnvironmentHThorClusterNames(StringArray &eclAgentNames, StringArray &groupNames, StringArray &targetNames)
-{
-    Owned<IEnvironmentFactory> factory = getEnvironmentFactory(true);
-    Owned<IConstEnvironment> env = factory->openEnvironment();
-    Owned<IPropertyTree> root = &env->getPTree();
-    Owned<IPropertyTreeIterator> allEclAgents = root->getElements("Software/EclAgentProcess");
-    ForEach(*allEclAgents)
-    {
-        IPropertyTree &eclAgent = allEclAgents->query();
-        const char *eclAgentName = eclAgent.queryProp("@name");
-        if (eclAgentName && *eclAgentName)
-        {
-            Owned<IPropertyTreeIterator> allTargets = root->getElements("Software/Topology/Cluster");
-            ForEach(*allTargets)
-            {
-                IPropertyTree &target = allTargets->query();
-                const char *targetName = target.queryProp("@name");
-                if (targetName && *targetName)
-                {
-                    StringBuffer xpath;
-                    xpath.appendf("EclAgentProcess[@process=\"%s\"]", eclAgentName);
-                    if (target.hasProp(xpath) && !target.hasProp("ThorCluster"))
-                    {
-                        StringBuffer groupName("hthor__");
-                        groupName.append(eclAgentName);
-
-                        groupNames.append(groupName);
-                        eclAgentNames.append(eclAgentName);
-                        targetNames.append(targetName);
-                    }
-                }
-            }
-        }
-    }
-    return eclAgentNames.ordinality();
-}
-
 
 IStringVal& CLocalWorkUnit::getScope(IStringVal &str) const 
 {
